requirements.txt
Django==4.2.16
psycopg2-binary
python-dotenv
openpyxl
django-jazzmin==3.0.1
reportlab
xlsxwriter
requests
django-cors-headers
msal
cryptography
